Fix typo and change loginStatus style and logout
