+++ frontend/src/core/services/PocketUserService.js
@@ -197,7 +197,7 @@ class PocketUserService extends PocketBaseService {
       email: this.getUserInfo().email,
     };
 
-    this.post(this._getURL("Auth/logout"), data)
+    axios.post(this._getURL("Auth/logout"), data)
       .then(logoutResponse => {
         if (logoutResponse.status === 200 && logoutResponse.data) {
           this.removeUserFromCached();
